Méthodes supplémentaire => Service Player

+++ src/main/java/aaa/tavern/service/PlayerService.java
@@ -2,9 +2,9 @@ package aaa.tavern.service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.DataAccessException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
@@ -27,7 +27,6 @@ public class PlayerService {
 
     // retourne l'id du player créer
     public void createPlayer(PlayerDto playerDto) {
-        //TODO avoir avec Adrien
             Role playerRole = roleRepository.findByName("USER").get() ;
             List<Role> roles = new ArrayList<Role>();
             roles.add(playerRole);
@@ -40,13 +39,69 @@ public class PlayerService {
                 ) ;
             playerRepository.save(newPlayer);
     }
+    // TODO méthodes de modification du Player sensible aux hacks
+    public boolean changeNickname(int idPlayer, String nickname){
+        Optional<Player> optPlayer =  playerRepository.findById(idPlayer) ;
+        if(optPlayer.isEmpty()){
+            return false ;
+        }else{
+            optPlayer.get().setNickname(nickname);
+            playerRepository.save(optPlayer.get()) ;
+            return true ;
+        }
+    }
+
+    public boolean changeEmail(int idPlayer, String email){
+        Optional<Player> optPlayer =  playerRepository.findById(idPlayer) ;
+        if(optPlayer.isEmpty()){
+            return false ;
+        }else{
+            optPlayer.get().setNickname(email);
+            playerRepository.save(optPlayer.get()) ;
+            return true ;
+        }
+    }
+
+    public boolean changePassword(int idPlayer, String password){
+        Optional<Player> optPlayer =  playerRepository.findById(idPlayer) ;
+        if(optPlayer.isEmpty()){
+            return false ;
+        }else{
+            optPlayer.get().setPassword(passwordEncoder.encode(password));
+            playerRepository.save(optPlayer.get()) ;
+            return true ;
+        }
+    }
 
     public boolean deletePlayer(int idPlayer) {
-        try {
+        Optional<Player> optPlayer =  playerRepository.findById(idPlayer) ;
+        if(optPlayer.isEmpty()){
+            return false ;
+        }else{
             playerRepository.deleteById(idPlayer);
             return true;
-        } catch (DataAccessException e) {
-            return false;
+        }
+    }
+
+    public boolean enabledPlayer(int idPlayer){
+        Optional<Player> optPlayer =  playerRepository.findById(idPlayer) ;
+        if(optPlayer.isEmpty()){
+            return false ;
+        }else{
+            optPlayer.get().setEnabled(true);
+            playerRepository.save(optPlayer.get()) ;
+            return true ;
+        }
+    }
+
+    public boolean disabledPlayer(int idPlayer){
+        Optional<Player> optPlayer =  playerRepository.findById(idPlayer) ;
+        if(optPlayer.isEmpty()){
+            return false ;
+        }else{
+            optPlayer.get().setEnabled(false);
+            playerRepository.save(optPlayer.get()) ;
+            return true ;
         }
     }
     /*
